// frontend/apps/stitch-admin/src/shared/api/generated/model/destinationResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Stitch API Gateway
 * OpenAPI spec version: 1.0.0
 */
import type { AddressDTO } from './addressDTO';
import type { DestinationType } from './destinationType';

export interface DestinationResponse {
  id: string;
  tenant_id: string;
  name: string;
  type: DestinationType;
  address: AddressDTO;
  is_active: boolean;
  contact_name: string | null;
  contact_phone: string | null;
}